# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open API documentation for project Album-singers",
        "contact": {
            "email": "galina@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
        "/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "albumsAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Create yet another album record",
                "operationId": "albumCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name album",
                                        "type": "string",
                                        "example": "Example"
                                    },
                                    "singer_id": {
                                        "description": "Id by singer",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "year": {
                                        "description": "Year album",
                                        "type": "integer",
                                        "example": "2022"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created siccessful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "Name album",
                                            "type": "string",
                                            "example": "Example"
                                        },
                                        "singer": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SingerShowResource"
                                            }
                                        },
                                        "year": {
                                            "description": "Year album",
                                            "type": "integer",
                                            "example": "2022"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "albumsOne",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of album",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Siccessful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "Album name",
                                            "type": "string",
                                            "example": "Example"
                                        },
                                        "singer": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SingerShowResource"
                                            }
                                        },
                                        "year": {
                                            "description": "Album year",
                                            "type": "integer",
                                            "example": "2022"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Delete album by ID",
                "operationId": "albumsDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of album",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successful"
                    },
                    "404": {
                        "description": "Model not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Albums"
                ],
                "summary": "Update yet another album record",
                "operationId": "albumUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of album",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Album name",
                                        "type": "string",
                                        "example": "Example"
                                    },
                                    "singer_id": {
                                        "description": "Singer ID",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "year": {
                                        "description": "Album year",
                                        "type": "integer",
                                        "example": "2022"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created siccessful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "Album name",
                                            "type": "string",
                                            "example": "Example"
                                        },
                                        "singer": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SingerShowResource"
                                            }
                                        },
                                        "year": {
                                            "description": "Album year",
                                            "type": "integer",
                                            "example": "2022"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Model not found"
                    }
                }
            }
        },
        "/albums/{id}/storeSongList": {
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Attach SongList",
                "operationId": "storeSongList",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of album",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "songs": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "song_id": {
                                                    "description": "Singer ID",
                                                    "type": "integer",
                                                    "example": "5"
                                                },
                                                "place": {
                                                    "description": "Song place",
                                                    "type": "integer",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SongList attached siccessful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SingerShowResource": {
                "title": "Singer model",
                "description": "Some simple request createa as example",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "5"
                    },
                    "name": {
                        "description": "The name",
                        "type": "string",
                        "example": "Alex"
                    }
                },
                "type": "object"
            }
        }
    },
...
}